chown all contents of the nix store to root:root
Initial motivation for this change:
Azure image created on an Ubuntu system with nix installed via curl to
sh method shown on the website creates a nix store in the image that
belongs to the uid of the user running nix on the Ubuntu host. This
causes at least sudo to fail as it complains that some file does not
belong to root.