gajim: Add gnome3.defaultIconTheme to buildInputs
This is for systems that are not using a full desktop environment, so
Gajim can fall back to the default icon theme of Gnome 3.

Among just fixing aesthetics this also fixes a few exceptions that were
triggered by Gajim not finding the icons. One example of this is when
you enter the plugins dialog, the window is empty if the icons can't be
found.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @abbradar, @Mic92, @7c6f434c
(cherry picked from commit a0c525bd0a93fba70471392b8651882b52197788)