nextcloud: add phpPackages config option
I would rather just take this from phpfpm service, but there is no such
attribute.  This ought to be configurable since it will be used for the caching
modules, and it has to match the version of php given to phpfpm.

(cherry picked from commit 716dcc4e6394df75915a0d621c12521b510eb669)