nixos/meshcentral: allow injecting arbitrary secret values
Required for OIDC support where you have to specify a `clientsecret` in
the `meshcentral-config.json`[1] which really shouldn't end up in the
store.

However I got tired of these hacks where you specify

    services.snens.settings = {
      secret = "$SECRET";
    };

because the entire section is _entirely meaningless_.

I figured it's nicer to inject the secrets as they are used in the
config, i.e.

    services.meshcentral.secrets.domains."".authStrategies.oidc.clientsecret = "/run/secrets/sekret";

loads the file `/run/secrets/sekret` via `LoadCredential` into
`meshcentral.service` and before startup its contents will be
injected into the JSON structure.

In the longterm it's perhaps even nicer to merge `secrets` and
`settings`, but given that I had a use-case for this now and I
dislike the "traditional" approach, I decided to try this idea out
now.

[1] https://ylianst.github.io/MeshCentral/meshcentral/config/