mailmanPackages: remove psycopg2 pin
This hack was necessary for django 2.x[1] which was the default `django`
version (i.e. current LTS) when this fix was needed. However, 2.x is now
EOL and not packaged anymore and mailman is running with django3, so
this can be dropped.

[1] https://github.com/psycopg/psycopg2/issues/1293

(cherry picked from commit 9a73e176acacf390f61eb6f1ba4bd71907716b4b)
Signed-off-by: Christian Albrecht <christian.albrecht@mayflower.de>